Move RGBQUAD struct to Gdi32
